LPS-57796 Move to its own class
