+++ modules/portal/portal-portlet-container-test/test/integration/com/liferay/portal/portlet/container/test/LayoutRequestPortletContainerTest.java
@@ -1,0 +1,59 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.portlet.container.test;
+
+import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
+import com.liferay.portal.kernel.test.rule.AggregateTestRule;
+import com.liferay.portal.kernel.test.rule.TransactionalTestRule;
+import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
+
+import java.util.List;
+import java.util.Map;
+
+import javax.servlet.http.HttpServletRequest;
+
+import org.junit.Assert;
+import org.junit.ClassRule;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+/**
+ * @author Raymond Augé
+ */
+@RunWith(Arquillian.class)
+public class LayoutRequestPortletContainerTest
+	extends BasePortletContainerTestCase {
+
+	@ClassRule
+	@Rule
+	public static final AggregateTestRule aggregateTestRule =
+		new AggregateTestRule(
+			new LiferayIntegrationTestRule(), TransactionalTestRule.INSTANCE);
+
+	@Test
+	public void testLayoutRequest() throws Exception {
+		setUpPortlet(testPortlet, properties, TEST_PORTLET_ID);
+
+		HttpServletRequest request = getHttpServletRequest();
+
+		Map<String, List<String>> responseMap = request(
+			layout.getRegularURL(request));
+
+		Assert.assertEquals("200", getString(responseMap, "code"));
+		Assert.assertTrue(map.containsKey("render"));
+	}
+
+}
